power-levels: Fix Total rooms summary output
Actually "Total rooms existing on server" was wrong: Since we use the
room-list-admin-API and thus can filter using the -n "search term"
option it's obviously that not always all rooms are shown. It's rather a
"total" of everything being found. Let's keep that as in the original
room list command output and just state: "Total Rooms"