docs: limit header links to fix extra height
if the site header is too tall, it interferes with pretty much all
elements of the site, slightly or very noticeably (e.g. when skipping to
headers in pages)

`navbar_align` being left is important at 960px width, because it just
so happens to not cause that (at least with inclusive sans). this is
tricky to get right though, opendyslexic needs
header_links_before_dropdown to be 1 so it doesn't cause an overflow

also hide the last link behind a "more" button to also reduce space
taken up